Move gids validation to GermplasmValidator
Issue: IBP-3750
