+++ src/main/java/org/ibp/api/java/impl/middleware/common/validator/GermplasmValidator.java
@@ -1,16 +1,27 @@
 package org.ibp.api.java.impl.middleware.common.validator;
 
+import org.generationcp.middleware.manager.api.GermplasmDataManager;
+import org.generationcp.middleware.pojos.Germplasm;
+import org.ibp.api.Util;
 import org.ibp.api.domain.germplasm.GermplasmSummary;
+import org.ibp.api.exception.ApiRequestValidationException;
 import org.ibp.api.java.germplasm.GermplasmService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.BindingResult;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Component
 public class GermplasmValidator {
 
 	@Autowired
 	private GermplasmService germplasmService;
+
+	@Autowired
+	private GermplasmDataManager germplasmDataManager;
 
 	public void validateGermplasmId(final BindingResult errors, final Integer germplasmId) {
 		if (germplasmId == null) {
@@ -20,6 +31,18 @@ public class GermplasmValidator {
 		final GermplasmSummary germplasmSummary = germplasmService.getGermplasm(String.valueOf(germplasmId));
 		if (germplasmSummary == null) {
 			errors.reject("germplasm.invalid", "");
+		}
+	}
+
+	public void validateGids(final BindingResult errors, final List<Integer> gids) {
+		final List<Germplasm> existingGermplasms = germplasmDataManager.getGermplasms(gids);
+		if (existingGermplasms.size() != gids.size() || existingGermplasms.stream().filter(g -> g.getDeleted()).count() > 0) {
+			final List<Integer> existingGids =
+				existingGermplasms.stream().filter(g -> !g.getDeleted()).map(Germplasm::getGid).collect(Collectors.toList());
+			final List<Integer> invalidGids = new ArrayList<>(gids);
+			invalidGids.removeAll(existingGids);
+			errors.reject("gids.invalid", new String[] {Util.buildErrorMessageFromList(invalidGids, 3)}, "");
+			throw new ApiRequestValidationException(errors.getAllErrors());
 		}
 	}
