+++ src/main/resources/messages_en.properties
@@ -390,6 +390,7 @@ inventory.unit.required=Inventory unit required
 inventory.unit.invalid=Invalid inventory unit identifier
 germplasm.required=Germplasm is required
 germplasm.invalid=Invalid germplasm identifier
+gids.invalid=The following gid(s): {0} were not found in the system
 
 transaction.notes.length=The max length of Transaction Notes is 255 characters
 transaction.wrong.transaction.status=The following transactions {0} are not in Pending Status. This action is only allowed for Pending Transactions. Please review.
@@ -409,7 +410,6 @@ lot.input.list.null=List of lots can not be null
 lot.input.list.no.items=List of lots to import is empty. Please review and specify at least one lot to import
 lot.input.list.item.null=Some of the elements in the lists are null
 lot.input.list.gid.null=GIDs are mandatory, please check for empty values
-lot.input.invalid.gids=The following gid(s): {0} were not found in the system
 lot.input.list.location.abbreviation.null.or.empty=Storage location abbreviations are mandatory, please check for empty values
 lot.input.invalid.abbreviations=The following Storage Location ABBR(s): {0} does not exist or not correspond to a Storage Location in the system. Please correct them and try again.
 lot.input.list.units.null.or.empty=Units are mandatory, please check for empty values
